file_info_hash_insert_name_size(): include file basename as an additional alias.
The filename is not part of the persisted aliases in the trailer, and it is
also not in the ALIA section in the ASCII database (it is listed under NAME).

Therefore, the actual file name was never considered for inclusion in the
table recording name+size entries.  This is OK as long as all persisted magnets
bear the SHA1 of the file, but if they miss it, we would not be able to tie
back the source to the proper fileinfo, creating another one each time!