forget_hashlist_node(): force node status to KNODE_UNKNOWN, unconditionally.
When the node is leaving the routing table, it must be flagged as such to
prevent further assertion failures when its refcount reaches 0 (a node in the
routing table could not reach this status unless there was a refcount bug).

The downside of that change is that we now need to forcefully reset the
status of the nodes in the split/merged buckets according to the list they
belong to, once the original bucket being split/merged has been cleared.

However, this will prevent subtle assertion failures when the nodes referred
to by the split/merged buckets are also part of an active DHT lookup.