dh_route(): only compress UDP query hits when GUESS queryier supports them.
This was not a bug so far, because the only GUESS querying party on the
network was gtk-gnutella, which supports the "Gnutella UDP Traffic Compression"
extension.

However, this was a shortcut made which was not future-proof, as other Gnutella
clients doing GUESS queries may appear and they may not support UDP traffic
compression.