QRP compression task now explicitly references the task it puts to sleep.
The BG task layer was enhanced to handle reference counting on tasks.

That new feature is primarily used in the compression task, which references
another task that it puts to sleep and then later on wakes up.  To prevent
problems when that other task is cancelled by the user code, it is necessary
to explicitly mention that we are taking an extra reference on the task,
and then later to explicitly remove that reference when we're no longer
going to access the task.