page_cache_insert_pages(): fixed incorrect write-lock upgrade logic.
When atomically upgrading from read to write lock, we were not setting the
wlock flag to TRUE, meaning we incorrectly released the lock as a read-lock
at the end of the routine.

However, when the atomic upgrade of read to write lock failed and we had to
release the read lock before requesting the write lock, things were working
as expected.