Allow for nested lock waiting sequences.
When a thread has registered it is waiting on a lock, and is therefore
"sleeping" until it is available, we could invoke deadlock tracing code
that can in turn need to sleep on (hopefully) other locks, creating a nested
waiting sequence.

Although this is a rare event, it can nonetheless happen and we need to prepare
for it or we can have inconsistent thread lock dumps (and more specifically
incorrect dumping of the locks being waited for) should the application
really end-up being deadlocked!