mingw_exception(): write shorter messages in case of stack overflows.
Instead of building a larger line describing what the problem is, emit
shorter messages with fewer information, in the hope that it will consume
less resources and we will be able to gather more output finally.

Also, don't terminate the calling thread (the one receiving the exception)
by calling thread_exit() when we are dealing with a discovered thread, i.e.
one that we have not created ourselves.