Added logging of layer which reports a TX error in case of assert failure.
When tx_has_error() is set and yet a write() operation on the queue is not
returning -1, we have a bug because all writes should either succeed or report
an error, but it's not possible to report success and yet have an error in the
TX stack.

To help diagnose which layer could be mistaken, we now report in the assertion
failure the name of the first layer that has an error condition flagged.