Added hook to actively monitor exiting of discovered threads.
Because GTK2 now creates threads underneath to manage things like the
"file selector" window (and not just one thread!) it is important to be
able to manage the threads correctly; otherwise we will overflow our thread
table which is limited to 64 entries when the "file selector" is being
extensively used in a session!

So now we can be warned when a discovered thread is exiting and reclaim
its thread descriptor appropriately after some grace period, provided we
do not witness any more activity for that thread.

Also revisited the stack overlapping logic of thread_element_tie() which
was not properly handling all the possible configurations.