thread_element_tie(): check for "exit_started", not "gone" (only set later).
Discovered threads that are going to die are know to run when the
thread_will_exit() hook is called, since it is done synchronously.

At that time, we set the "exit_started" flag in the thread element, but
the "gone" flag will only be set later, a after some cleanup grace time
to avoid mistaking cleanup activity with the thread actually being alive.

Therefore, we need to handle the "exit_startup" flag otherwise we would run
the risk of marking a thread element reusable before the "gone" flag is
set, leading to that thread element being reused and then all the processing
chain for that gone thread will be wrong if that element was reused by a
created thread for instance.