pmap_insert_region(): must recompute insertion index if we extend the pmap!
At the end of the insertion procedure, we extend the pmap if it became full.
However, this means we must recompute the index since the pages that constitute
the pmap are themselves part of the pmap. Hence the actual index where the
region we wanted to insert with pmap_insert_region() has been stored could have
changed.