bg_info_get(): prevent possible deadlock by not always locking task.
Since we're mostly reading, we can get a lot of the information we want
without really holding the lock.  Whether the information was gathered
under lock is also reported so that the user layer can flag a possible
inconsistency.

The potential for deadlock was spotted by Coverity under issue CID 106476.