qrt_compressed(): was not handling QRP compression task cancellation properly.
What could happen was the following call chain:

    mq_tcp_putq
    gmsg_sendto_one
    qrt_compressed
    bg_task_finished
    bg_task_cancel
    qrt_update_free
    node_remove_v
    node_tx_eof_remove
    tx_link_write_error
    tx_link_write
    tx_write

Reading the stack from bottom up: on a link write error, the TX stack was
reporting the error invoking the user-supplied callback: removing the faulty
node, cancelling any pending QRP compression task.  Normal processing.

But due to the bug in qrt_compressed(), we were not seeing the task had been
cancelled and went on sending a reset to the faulty node, re-entering the
message queue.  Eventually the write would succeed because it would end-up
having its data buffered by the deflate layer, and we would trigger an
assertion like:

	Assertion failure at src/core/mq_tcp.c:385:
		"(ssize_t) -1 == written || !tx_has_error(q->tx_drv)"

	mq_tcp_putq(): written=29, yet error reported by TX layer "link"

In this fix, we restore proper handling of the cancellation by the bg task
completion callback, to avoid writing anything back to the message queue,
and therefore prevent a crash.