Added crash_restart() and controlling routines to request smooth restart.
This is now used by our Windows VMM low-level page allocator when it detects
that the application is running low on unreserved space.  Instead of a brutal
restart, this allows the context to be saved completely and cleanly before
letting the application relaunch itself.

This relies on the fact that we're detecting the potential out-of-memory
condition well before it will actually happen.