On network buffer shortage (ENOBUFS), stop issuing TCP connections for a while.
This seems to happen quite frequently on Windows.  Since it is not clear where
the error comes from, we suppose that there is a temporary shortage of IP
resources and avoid issuing TCP connections (which are heavy consumers of
buffer resources) for a while, hoping the condition will clear itself in time.