Cleaned up crash_restart() and its configuration.
If they have not configured the application to auto-restart, crash_restart()
is now silently ignored.

The crash_set_restart() routine will also loudly warn when it is called
before crash_init() since it cannot record the restart callback when the
structure used to store that callback has not been allocated yet.