node_add_internal(): for outband connections, flag Gnutella or G2 hosts.
When we are connecting to a remote host, we know whether we are targetting
an Ultrapeer (for Gnutella) or a Hub (for G2), so make that appear in the
node flags.  Later on, if we end-up connecting to a leaf that accepts us,
the 'U' flag will correctly turn into a 'L'.

This lets users see which network the outbound connections are aiming for.