Loudly warn if attempting a file-descriptor operation on a revoked descriptor.
When a file object is told to unlink a file, we are revoking the descriptor.
Any further attempt to interact with the file object to fstat(), ftruncate()
it or get its file descriptor will now result in a loud warning instead of an
assertion failure.

The errno is properly set when needed so that upper layers can trap the error.
Still, the loud warning is there because there is a logic error in the
application, resulting in an invalid use of file objects.  Nothing critical
warranting an immediate shutdown though as no corruption is possible.