page_cache_find_pages(): treat user memory as if under short-term strategy.
User memory is less likely to be long-lived as it will typically be allocated,
used, and then freed as a whole.  Therefore, it is more suitable to handle
these allocation as if they were made under a short-term strategy, to make
sure we can serve them out of the page cache, even for single pages.