Added workaround for Phex: always send X-Gnutella-Content-URN in upload replies.
Normally, gtk-gnutella never sends that line on a follow-up request because it
already advertised the proper URN on the previous request.  Also, when bandwith
is saturated, we avoid sending back this header if they requested a file through
an N2R request, where they had to supply the SHA1 anyway.

But Phex expects this header, always, or it will break up the connection.

Since Phex is no longer maintained, but there are still Phex servents out there,
we work-around that "bug" in Phex by always sending back the expected header
when the request comes from a Phex servent.