qrt_patch_compress_done(): use new "finished" flag to signal task is completed.
There was a corner-case happening sometimes in qrp_comp_done(): when the
query routine table compress was not flagged as "done", it was attempting
to cancel it, then expecting to have "done" set since cancellation was meant
to be synchronous.

Problems arise for two reasons:
1) The "done" flag was actually set in the compress free routine.
2) The "completion callback on the patch compression task was not seting
   the "done" flag.

So we're introducing a new "finished" flag that is now set in the completion
callback routine, qrt_patch_compress_done(), and this is what qrp_comp_done()
checks for and expects to be set when it has to cancel the compression task.