pmap_remove_from(): protect from exceptional corner case to avoid crash.
When extending a full pmap, the new allocated pmap may be partially overlapping
a foreign mapping.  During overruling of the allocated segment, we can attempt
to remove an inner part of a foreign mapping.  This creates a new mapping, that
we cannot insert since the pmap is full and being extended.

In that case, we simply drop the foreign mapping information, which is fine
since that information is intuited from failed allocations and could be either
wrong or stale.