EVQ: closed race condition betweeen queues recreated in same thread ID.
A thread "n" could have an event queue with event "E" enqueued to it.
The event queue thread would schedule "E", but then would be suspended
by the scheduler and during that time, thread "n" dies, a new thread is
created with the same ID with a queue installed, and then the event queue
thread is resumed... presenting the obsolete event "E" to evq_trampoline().