omalloc_allocate(): was releasing spinlock too soon in read-only mode.
When allocating from a read-only chunk, we must keep the critical section until
we have restored the read-only permission on the pages where the allocation
was made.

Failure to do so opens a race-condition whereby the chunk is protected back
to read-only mode whilst another thread already thinks it has made it
read-write to be able to allocate from it, causing a memory fault in that
other thread.