Added spopen(), a simple popen() bypassing the shell.
The reasons we might need the spopen() family of functions are:

- spopen() bypasses the shell, meaning arguments are not interpreted,
  causing no risk with unescaped meta-characters present in the arguments.

- spopen() returns a kernel file descriptor, not a FILE *, hence there is
  no buffering when writing down to the pipe (but the child might buffer
  its output to the parent process).

- spopen() works on Windows.

The downside of not having the shell interpret the arguments is that the
signature of spopen() needs to account for possible redirections in the
child process.  This is a minor problem and the interface is simple enough
(a single fd[] array with 2 items), yet expressive enough to cover most of
the redirection needs.