SDBM: enhanced loose iterations by adding a DBM_F_ALLKEYS flag.
Sometimes a loose iteration needs to be predictible to make sure we do not
unduly skip a key on a page or ensure that all delete operations will be
able to occcur.

The DBM_F_ALLKEYS flag ensures that the database is locked when each page
is processed, avoiding the risk of concurrent updates on the page at the
price of keeping the database lock for a longer period of time (since all
the callbacks are invoked with the lock being held).