Added Windows-specific PROT_GUARD protection type for mprotect().
On UNIX machines, this is transparently mapped to PROT_NONE since there is
no equivalent features offered.  Of course, the mechanism could be emulated
via a SIGSEGV signal handler and some bookkeeping of memory region protections.

The reason PROT_GUARD is necessary on Windows is that there is no sigaltstack()
system call to request that SIGSEGV signals be sent on an alternate execution
stack.  There are no SIGSEGV signals either, but there are exception handlers.