Windows: reworked stack overflow detection logic.
On Windows, the kernel reserves up to 2 MiB of stack space in the VM, marking
the end with a guard page so that it can dynamically grow the stack as needed
to meet application demand.

However, we usually want to limit stack space to be able to detect wrong
application behaviour.  On UNIX platforms where we can direct pthreads to
allocate the stack for us at the right size, this allows us to also avoid
eating up the VM space, especially on 32-bit platforms.

On Windows unfortunately, no stack overflow will be detected and reported
actively until we hit the 2 MiB barrier of stack consumption.  At that time
there is only a limited amount of stack, too small to run the full exception
reporting, stack backtracing and possibly the thread signal handler for the
TSIG_OVFLOW signal that our runtime generates.

On top of that, there is no sigaltstack() on Windows.

Therefore, our new strategy is to monitor the actual stack consumption, which
is going to always be smaller than the 2 MiB limit imposed by the system.  As
soon as we detect that the application uses more stack, we trigger the stack
overflow logic and we know we have sufficient stack space available to actually
perform all the cleanup and logging necessary.