By default, gtk-gnutella now supervises its children to restart them.
When --no-supervise is given, we revert to the old behaviour of running
directly as a standalone program.

Otherwise, a parent process will monitor children processes and relaunch them
if they are exiting abnormally.   Children make sure to not auto-restart when
they are supervised, unless their parent disappears for some reason.