xmalloc: re-introduced xpmalloc() to request allocation from free lists.
The xprealloc() call is there to reallocate such blocks.

The aim is to be able to ensure there is a header in the allocated block,
which is not the case when allocating from a thread memory pool.  It is
necessary for halloc() when the maximum size available for walloc() was
explicitly reduced.