Refactored main() initializations to limit memory usage in the supervisor.
When the supervisor process initializes, it disables halloc() and limits
walloc() so that most memory allocations be done through xmalloc().

Given that the tasks the supervisor has to do are very simple, there is no
need to allocate much memory.  Hence there is no need to create extra data
structures and setup large zones for optimal small block allocations!