Added specific PID file for locking out the supervisor process.
Since the supervisor process is the one that is launched by default, we
need to ensure there are no two such processes running.

If we did not, the supervisor would fail anyway because all the children it
tried to launch would quickly exit when they detected that another gtk-gnutella
was running, but it would take a few seconds and the reason would not be as
obvious.  Better emit the proper diagnostic right away!