unique_file_token(): include the hostname into the unique file ID.
This protects against duplicate GUID or KUID being used when the file resides
on a virtual machine which has been cloned.  The device and inode are going
to be identical, but the hostname of the cloned machine will likely be
different.

This change is going to cause all the property files out there to be handled
as being "stale", so all the GUIDs and KUIDs will be recomputed.