Revisited crash_mode() to include the notion of criticity level.
Actions that are taken are now based on the level of the crash, so that we
do not disable useful things all the time.  For instance, in out-of-memory
conditions,  we can avoid disabling memory allocators, which may have some
internal room available for small allocations, and thus peruse them, as
opposed to stopping all allocation / freeing and starting to consume the
emergency memory.