mingw_execve(): redesigned to avoid using external resources.
Indeed, during out-of-memory conditions, we want to be in control of our
memory allocation, to benefit from the VMEA reserve.  Hence we cannot make
use of the spawnve() routine which will allocate memory of its own and could
fail when we are precisely out of memory!

Also, we must make sure to release stdout and stderr or the new process will
not be able to properly rename the file and we will lose all the information
that was stored there, not able to figure out what went wrong and led to a
restart.