Added THREAD_F_CLEARSIG creation flag to prevent signal mask inheritance.
All the automatically triggered thread creation (time, callout queue
and event queue) now make use of that flag to ensure they will accept
any thread signal, regardless of the state of the signal mask of the
thread that created them.

Since we do not control the auto-initialization sequence, which can
vary depending on the platforms, this is a safe way to ensure a control
of the thread initial mask.  If a thread needs to further tailor its
mask to avoid certain signals, it can always do so via thread_sigmask()
at startup.