Suspend threads created after a global suspension is in effect.
Unless they give the THREAD_F_UNSUSPEND flag at creation time, all the
threads that were in a creation state but not yet launched will automatically
suspend themselves before entering their main routine when there is a global
thread suspension in effect.

A global suspension is usually there to avoid side effects when crashing.
However, on Windows, the crash handler needs to start a thread to launch
the debugger, hence the need for THREAD_F_UNSUSPEND to still permit the
thread to proceed despite the probable global suspension at that time!