thread_element_update_qid_range(): avoid soft assertion, use warning instead.
During crashes, when the soft assertion was failing, a deadly recursion was
created through assertion_message() which called thread_check_suspended()
which again re-entered thread_element_update_qid_range().

With a warning message, there is no danger of recursion, and we can log
outside the critical section on the thread element.