rwlock: no longer block signals when waiting for a rw lock.
The new strategy to prevent deadlocks is now to explicitly check whether we
are running in a signal handler.  When we do, we are very careful and only
append ourselves to the waiting list provided we do not find ourselves already
listed among the waiters.  If we do find ourselves, we pre-empt our execution
by prepending the signal handler execution thread to the existing record.

This means thread_sighandler_level() must no longer check for pending signals
since we do not want a signal to be triggered whilst we are enqueuing ourselves
in the queue and call thread_sighandler_level() to see whether we need to be
careful!