stacktrace_unwind(): use gcc-style unwinding if recursing or within malloc.
In case we are running from a signal handler, check whether we are coming
from malloc, by looking at whether we are holding any lock there.  This is
far from perfect, but it should allow to use backtrace() in a relatively
safe way -- we know xmalloc() will not use a thread-specific pool when coming
from a signal handler, hence critical sections will show up.

This is because backtrace() can potentially invoke malloc().